fix latenclies charts

+++ Microsoft.Azure.Cosmos.Samples/Tools/Benchmark/BenchmarkLatencyEventSource.cs
@@ -47,13 +47,13 @@ namespace CosmosBenchmark
         }
 
         [Event(2, Level = EventLevel.Informational)]
-        public void OnOperationSuccess(int operationType, long durationInMs)
+        public void OnOperationSuccess(int operationType, double durationInMs)
         {
             this.WriteEvent(2, operationType, durationInMs);
         }
 
         [Event(3, Level = EventLevel.Informational)]
-        public void OnOperationFailure(int operationType, long durationInMs)
+        public void OnOperationFailure(int operationType, double durationInMs)
         {
             this.WriteEvent(3, operationType, durationInMs);
         }
